pearch_digital_test_api/requirements.txt
Django
djangorestframework
gunicorn
drf-generators
django-typomatic
black
pyyml
uritemplate
django-rest-swagger
django-rest-auth
django-allauth
drf-api-logger
django-cors-headers
django-filter
drf-yasg
pytest-django